Add ModelStep class
This class defines a step for running the E3SM component as part of
a test case.

eventually, each step that currently calls `run_model()` should
instead descend from `ModelStep`
