+++ polaris/__init__.py
@@ -1,4 +1,5 @@
 from polaris.component import Component
+from polaris.model_step import ModelStep
 from polaris.step import Step
 from polaris.testcase import TestCase
 from polaris.testgroup import TestGroup
