+++ polaris/step.py
@@ -4,13 +4,9 @@ import os
 import shutil
 import stat
 
-import numpy
 import progressbar
-from lxml import etree
 from mache import MachineInfo
 
-import polaris.namelist
-import polaris.streams
 from polaris.config import PolarisConfigParser
 from polaris.io import download, package_path, symlink
 
@@ -432,218 +428,6 @@ class Step:
         """
         self.outputs.append(filename)
 
-    def add_model_as_input(self):
-        """
-        make a link to the model executable and add it to the inputs
-        """
-        self.add_input_file(filename='<<<model>>>')
-
-    def add_namelist_file(self, package, namelist, out_name=None,
-                          mode='forward'):
-        """
-        Add a file with updates to namelist options to the step to be parsed
-        when generating a complete namelist file if and when the step gets set
-        up.
-
-        Parameters
-        ----------
-        package : Package
-            The package name or module object that contains ``namelist``
-
-        namelist : str
-            The name of the namelist replacements file to read from
-
-        out_name : str, optional
-            The name of the namelist file to write out, ``namelist.<core>`` by
-            default
-
-        mode : {'init', 'forward'}, optional
-            The mode that the model will run in
-        """
-        if out_name is None:
-            out_name = f'namelist.{self.component.name}'
-
-        if out_name not in self.namelist_data:
-            self.namelist_data[out_name] = list()
-
-        namelist_list = self.namelist_data[out_name]
-
-        namelist_list.append(dict(package=package, namelist=namelist,
-                                  mode=mode))
-
-    def add_namelist_options(self, options, out_name=None, mode='forward'):
-        """
-        Add the namelist replacements to be parsed when generating a namelist
-        file if and when the step gets set up.
-
-        Parameters
-        ----------
-        options : dict
-            A dictionary of options and value to replace namelist options with
-            new values
-
-        out_name : str, optional
-            The name of the namelist file to write out, ``namelist.<core>`` by
-            default
-
-        mode : {'init', 'forward'}, optional
-            The mode that the model will run in
-        """
-        if out_name is None:
-            out_name = f'namelist.{self.component.name}'
-
-        if out_name not in self.namelist_data:
-            self.namelist_data[out_name] = list()
-
-        namelist_list = self.namelist_data[out_name]
-
-        namelist_list.append(dict(options=options, mode=mode))
-
-    def update_namelist_at_runtime(self, options, out_name=None):
-        """
-        Update an existing namelist file with additional options.  This would
-        typically be used for namelist options that are only known at runtime,
-        not during setup, typically those related to the number of nodes and
-        cores.
-
-        Parameters
-        ----------
-        options : dict
-            A dictionary of options and value to replace namelist options with
-            new values
-
-        out_name : str, optional
-            The name of the namelist file to write out, ``namelist.<core>`` by
-            default
-        """
-
-        if out_name is None:
-            out_name = f'namelist.{self.component.name}'
-
-        print(f'Warning: replacing namelist options in {out_name}')
-        for key, value in options.items():
-            print(f'{key} = {value}')
-
-        filename = os.path.join(self.work_dir, out_name)
-
-        namelist = polaris.namelist.ingest(filename)
-
-        namelist = polaris.namelist.replace(namelist, options)
-
-        polaris.namelist.write(namelist, filename)
-
-    def update_namelist_pio(self, out_name=None):
-        """
-        Modify the namelist so the number of PIO tasks and the stride between
-        them consistent with the number of nodes and cores (one PIO task per
-        node).
-
-        Parameters
-        ----------
-        out_name : str, optional
-            The name of the namelist file to write out, ``namelist.<core>`` by
-            default
-        """
-        config = self.config
-
-        cores = self.ntasks * self.cpus_per_task
-
-        if out_name is None:
-            out_name = f'namelist.{self.component.name}'
-
-        cores_per_node = config.getint('parallel', 'cores_per_node')
-
-        # update PIO tasks based on the machine settings and the available
-        # number or cores
-        pio_num_iotasks = int(numpy.ceil(cores / cores_per_node))
-        pio_stride = self.ntasks // pio_num_iotasks
-        if pio_stride > cores_per_node:
-            raise ValueError(f'Not enough nodes for the number of cores.  '
-                             f'cores: {cores}, cores per node: '
-                             f'{cores_per_node}')
-
-        replacements = {'config_pio_num_iotasks': f'{pio_num_iotasks}',
-                        'config_pio_stride': f'{pio_stride}'}
-
-        self.update_namelist_at_runtime(options=replacements,
-                                        out_name=out_name)
-
-    def add_streams_file(self, package, streams, template_replacements=None,
-                         out_name=None, mode='forward'):
-        """
-        Add a streams file to the step to be parsed when generating a complete
-        streams file if and when the step gets set up.
-
-        Parameters
-        ----------
-        package : Package
-            The package name or module object that contains the streams file
-
-        streams : str
-            The name of the streams file to read from
-
-        template_replacements : dict, optional
-            A dictionary of replacements, in which case ``streams`` must be a
-            Jinja2 template to be rendered with these replacements
-
-        out_name : str, optional
-            The name of the streams file to write out, ``streams.<core>`` by
-            default
-
-        mode : {'init', 'forward'}, optional
-            The mode that the model will run in
-        """
-        if out_name is None:
-            out_name = f'streams.{self.component.name}'
-
-        if out_name not in self.streams_data:
-            self.streams_data[out_name] = list()
-
-        self.streams_data[out_name].append(
-            dict(package=package, streams=streams,
-                 replacements=template_replacements, mode=mode))
-
-    def update_streams_at_runtime(self, package, streams,
-                                  template_replacements, out_name=None):
-        """
-        Update the streams files during the run phase of this step using the
-        given template and replacements.  This may be useful for updating
-        streams based on config options that a user may have changed before
-        running the step.
-
-        Parameters
-        ----------
-        package : Package
-            The package name or module object that contains the streams file
-
-        streams : str
-            The name of a Jinja2 template to be rendered with replacements
-
-        template_replacements : dict
-            A dictionary of replacements
-
-        out_name : str, optional
-            The name of the streams file to write out, ``streams.<core>`` by
-            default
-        """
-
-        if out_name is None:
-            out_name = f'streams.{self.component.name}'
-
-        if template_replacements is not None:
-            print(f'Warning: updating streams in {out_name} using the '
-                  f'following template and replacements:')
-            print(f'{package} {streams}')
-            for key, value in template_replacements.items():
-                print(f'{key} = {value}')
-
-        filename = os.path.join(self.work_dir, out_name)
-
-        tree = etree.parse(filename)
-        tree = polaris.streams.read(package, streams, tree=tree,
-                                    replacements=template_replacements)
-        polaris.streams.write(tree, filename)
-
     def process_inputs_and_outputs(self):
         """
         Process the inputs to and outputs from a step added with
@@ -694,9 +478,6 @@ class Step:
         self.outputs = [os.path.abspath(os.path.join(step_dir, filename)) for
                         filename in self.outputs]
 
-        self._generate_namelists()
-        self._generate_streams()
-
     @staticmethod
     def _process_input(entry, config, base_work_dir, component, step_dir):
         database_subdir = None
@@ -708,9 +489,6 @@ class Step:
         work_dir_target = entry['work_dir_target']
         package = entry['package']
         copy = entry['copy']
-
-        if filename == '<<<model>>>':
-            filename, target = Step._process_model(config, base_work_dir)
 
         if package is not None:
             if target is None:
@@ -763,127 +541,6 @@ class Step:
             input_file = filename
 
         return input_file, database_subdir
-
-    @staticmethod
-    def _process_model(config, base_work_dir):
-        model = config.get('executables', 'component')
-        filename = os.path.basename(model)
-        copy_executable = config.getboolean('setup', 'copy_executable')
-        if copy_executable:
-            # make a copy of the model executable, then link to that
-            mpas_subdir = os.path.basename(
-                config.get('paths', 'mpas_model'))
-            mpas_workdir = os.path.join(base_work_dir, mpas_subdir)
-            target = os.path.join(mpas_workdir, filename)
-            try:
-                os.makedirs(mpas_workdir)
-            except FileExistsError:
-                pass
-
-            try:
-                shutil.copy(model, target)
-            except FileExistsError:
-                pass
-        else:
-            target = os.path.abspath(model)
-        return filename, target
-
-    def _generate_namelists(self):
-        """
-        Writes out a namelist file in the work directory with new values given
-        by parsing the files and dictionaries in the step's ``namelist_data``.
-        """
-
-        if self.cached:
-            # no need for namelists
-            return
-
-        step_work_dir = self.work_dir
-        config = self.config
-
-        for out_name in self.namelist_data:
-
-            replacements = dict()
-
-            mode = None
-
-            for entry in self.namelist_data[out_name]:
-                if mode is None:
-                    mode = entry['mode']
-                else:
-                    assert mode == entry['mode']
-                if 'options' in entry:
-                    # this is a dictionary of replacement namelist options
-                    options = entry['options']
-                else:
-                    options = polaris.namelist.parse_replacements(
-                        entry['package'], entry['namelist'])
-                replacements.update(options)
-
-            defaults_filename = config.get('namelists', mode)
-            out_filename = f'{step_work_dir}/{out_name}'
-
-            namelist = polaris.namelist.ingest(defaults_filename)
-
-            namelist = polaris.namelist.replace(namelist, replacements)
-
-            polaris.namelist.write(namelist, out_filename)
-
-    def _generate_streams(self):
-        """
-        Writes out a streams file in the work directory with new values given
-        by parsing the files and dictionaries in the step's ``streams_data``.
-        """
-        if self.cached:
-            # no need for streams
-            return
-
-        step_work_dir = self.work_dir
-        config = self.config
-
-        for out_name in self.streams_data:
-
-            # generate the streams file
-            tree = None
-
-            mode = None
-
-            for entry in self.streams_data[out_name]:
-                if mode is None:
-                    mode = entry['mode']
-                else:
-                    assert mode == entry['mode']
-
-                tree = polaris.streams.read(
-                    package=entry['package'],
-                    streams_filename=entry['streams'],
-                    replacements=entry['replacements'], tree=tree)
-
-            if tree is None:
-                raise ValueError('No streams were added to the streams file.')
-
-            defaults_filename = config.get('streams', mode)
-            out_filename = f'{step_work_dir}/{out_name}'
-
-            defaults_tree = etree.parse(defaults_filename)
-
-            defaults = next(defaults_tree.iter('streams'))
-            streams = next(tree.iter('streams'))
-
-            for stream in streams:
-                polaris.streams.update_defaults(stream, defaults)
-
-            # remove any streams that aren't requested
-            for default in defaults:
-                found = False
-                for stream in streams:
-                    if stream.attrib['name'] == default.attrib['name']:
-                        found = True
-                        break
-                if not found:
-                    defaults.remove(default)
-
-            polaris.streams.write(defaults_tree, out_filename)
 
     def _fix_permissions(self, databases):  # noqa: C901
         """
